pyproject.toml
[project]
name = "ruetconnect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-cotton>=2.1.1",
    "django-htmx>=1.23.0",
    "django-tailwind-cli>=4.2.3",
    "django-template-partials>=24.4",
    "django-widget-tweaks>=1.5.0",
    "lucide[django]>=1.1.3",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "typer>=0.15.3",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.18.0",
    "django-tailwind-cli>=4.2.3",
    "flake8>=7.2.0",
    "pre-commit>=4.2.0",
]
prod = [
    "uvicorn>=0.34.2",
]
test = [
    "pytest>=8.3.5",
]


[tool.isort]
profile = "black"
line_length = 119
